Extend schema and code

+++ src/common/constants.ts
@@ -28,7 +28,7 @@ export const WETH_ADDRESS = Address.fromString(
 export const FACTORY_ADDRESS = '0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f'
 
 // minimum liquidity for price to get tracked
-export let MINIMUM_LIQUIDITY_THRESHOLD_ETH = BigDecimal.fromString('2')
+export let MINIMUM_LIQUIDITY_THRESHOLD_ETH = BigInt.fromString('2')
 
 export const ZERO_ADDRESS = Address.fromString(
   "0x0000000000000000000000000000000000000000"
